Ajustes envolvendo arquivos para teste
