+++ tests/test_csv_file.py
@@ -3,17 +3,6 @@ import os
 import insetos_hexapoda.csv_file as csv
 
 
-class Insect:
-
-    def __init__(self, genre: str,
-                 length: list[str],
-                 food: list[str]) -> None:
-
-        self.genre = genre
-        self.length = length
-        self.food = food
-
-
 def remove_file_temp(file_temp):
     os.remove(file_temp)
 
@@ -58,18 +47,5 @@ def test_passes_read_csv(setup):
     assert res
 
 
-def check_genre(mantodea: dict[str, Insect]) -> bool:
-
-    genre: str = mantodea["louva-a-deus"].genre
-
-    return "rhombodera" == genre.lower()
 
 
-def test_pass_check_genre():
-
-    insect = Insect("Rhombodera", ['12cm', '20cm'],
-                    ['rãs', 'lagartos', 'ratos', 'pássaros', 'pequenos', 'cobras'])
-
-    mantodea = {"louva-a-deus": insect}
-
-    assert check_genre(mantodea)
